projeto-db/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pymongo
neo4j
redis
python-dotenv
